+++ engine/move.hpp
@@ -183,7 +183,7 @@ class Move {
     static array<string, N> getUciArr(array<Move, N> arr) {
         array<string, N> uciArray;
 
-        for (int i = 0; i < arr.size(); i++) {
+        for (size_t i = 0; i < arr.size(); i++) {
             uciArray[i] = arr[i].toUci();
         }
 
@@ -194,7 +194,7 @@ class Move {
     static array<string, N> getUciArr(stackvector<Move, N> arr) {
         array<string, N> uciArray;
 
-        for (int i = 0; i < arr.size(); i++) {
+        for (size_t i = 0; i < arr.size(); i++) {
             uciArray[i] = arr[i].toUci();
         }
